==> 01_LLM_bib_BERTopic.ipynb <==
import pandas as pd
import numpy as np
import os

data_import_path = "/nfs/turbo/hrg/somar/LLM_bib/raw_data"
data_export_path = "/nfs/turbo/hrg/somar/LLM_bib/result_data"
result_path = "/nfs/turbo/hrg/somar/LLM_bib/result_file"

# # read the file into a pandas DataFrame
# wos_df = pd.read_csv(data_import_path+"/download_chatgpt_1001_1500.txt", delimiter='\t', header=0, index_col=False)

# # print the first 5 rows of the DataFrame
# print(wos_df.head())

wos_df = pd.DataFrame()
for file_name in sorted(os.listdir(data_import_path+"/excel_version")):
    print(file_name)
    if file_name.endswith(".xls"):
        wos_df = wos_df.append(pd.read_excel(data_import_path+"/excel_version/"+file_name))

wos_df.info()

wos_df = wos_df.reset_index(drop=True)

wos_abstract_df = wos_df[['Article Title','DOI','Abstract']].reset_index(drop=True)

wos_abstract_df["Article_number"] = list(range(5752))

# We use title and abstract, as well focus on sentence-line context aware info, not keywords.
wos_abstract_df["Title_abstract"] = wos_abstract_df["Article Title"]+". "+wos_abstract_df["Abstract"]

# add dates
all_years = []
for i in range(len(wos_df)):
    if not np.isnan(wos_df["Publication Year"][i]):
        all_years.append(int(wos_df["Publication Year"][i]))
    elif type(wos_df["Early Access Date"][i])==str:
        all_years.append(int(wos_df["Early Access Date"][i][-4:]))
    else:
        all_years.append("")

all_years.count("")

wos_abstract_df["Year"] = all_years

### Topic Modeling

from bertopic import BERTopic
from sentence_transformers import SentenceTransformer
from umap import UMAP
from sklearn.feature_extraction.text import CountVectorizer
from hdbscan import HDBSCAN
from sklearn.cluster import KMeans

docs = [str(item) for item in list(wos_abstract_df.Title_abstract)]

vectorizer_model = CountVectorizer(stop_words="english")
# hdbscan_model = HDBSCAN(min_cluster_size=10,
#                         min_samples=5,
#                         metric='euclidean',
#                         cluster_selection_method='eom',
#                         prediction_data=True) # random_state=42
cluster_model = KMeans(n_clusters=200)
topic_model = BERTopic(hdbscan_model=cluster_model, #hdbscan_model,
                       vectorizer_model=vectorizer_model,
                       top_n_words=10) # nr_topics=100,
topics, probs = topic_model.fit_transform(docs)

topic_model.save(result_path+"/LLM_wos_abstract_200")

wos_abstract_df["topics"] = topics

wos_abstract_df

# show topics info
topic_model.get_topic_info()

topic_model.get_topic_info().to_csv(data_export_path+"/LLM_wos_abstract_topics_metadata_200.csv",index=False)

# show Topic Positions -> suggestions for further clustering
#topic_model.visualize_topics()
fig_IDM = topic_model.visualize_topics()
fig_IDM.write_html(result_path+"/LLM_wos_abstract_viz_IDM_200.html")

# visualize Topic Hierarchy -> suggestions for further clustering
fig_hierarchy = topic_model.visualize_hierarchy()
fig_hierarchy.write_html(result_path+"/LLM_wos_abstract_viz_hierarchy_200.html")

# visualize words
fig_topic_word_scores = topic_model.visualize_barchart(top_n_topics=32)
fig_topic_word_scores.write_html(result_path+"/LLM_wos_abstract_viz_topic_word_scores_200.html")

# visualize topic similarity -- not useful in our case, maybe include to show not correlated
fig_heatmap = topic_model.visualize_heatmap()
fig_heatmap.write_html(result_path+"/LLM_wos_abstract_viz_heatmap_200.html")

# visualize documents
sentence_model = SentenceTransformer("all-MiniLM-L6-v2")
embeddings = sentence_model.encode(docs, show_progress_bar=False)
# Run the visualization with the original embeddings
topic_model.visualize_documents(docs, embeddings=embeddings)
# Reduce dimensionality of embeddings, this step is optional but much faster to perform iteratively:
reduced_embeddings = UMAP(n_neighbors=10, n_components=2, min_dist=0.0, metric='cosine').fit_transform(embeddings)
# save embeddings (for Tableau)
wos_abstract_df["x"] = reduced_embeddings[:, 0]
wos_abstract_df["y"] = reduced_embeddings[:, 1]


import pickle
with open(result_path+'/LLM_wos_abstract_embeddings_200.pickle', 'wb') as pkl:
    pickle.dump(embeddings, pkl)

wos_abstract_df.to_csv(data_export_path+"/LLM_wos_abstract_w_topics_years_positions_200.csv",index=False)

### Topic over time
topics_over_time = topic_model.topics_over_time(docs, all_years)

# will export tableau viz after finish
# topics_over_time_eg_viz = topic_model.visualize_topics_over_time(topics_over_time, topics=[13, 2, 8, 10, 40, 47])
# topics_over_time_eg_viz.write_html(result_path+"/LLM_wos_abstract_viz_topics_over_time_eg.html")

TODO:
1. Remove nan docs
2. Group the 50 topics, or choose interesting ones

Explore fields
1. get unique
2. coref
3. detect ROR
4. analyze collaboration

Refs:
- https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5065891/
- https://core.ac.uk/download/pdf/25068797.pdf

Note: we don't care about co-citation, we just aggregate collaboration in single publications

wos_df.Affiliations

# wos_df.Addresses[2]
