==> 02_LLM_Bib_Annotated_Viz.ipynb <==
import pandas as pd
import numpy as np

data_import_path = "/nfs/turbo/hrg/somar/LLM_bib/raw_data"
data_export_path = "/nfs/turbo/hrg/somar/LLM_bib/result_data"
result_path = "/nfs/turbo/hrg/somar/LLM_bib/result_file"

df = pd.read_csv(data_export_path+"/LLM_wos_abstract_w_topics_years_positions_200.csv")
ann = pd.read_csv(data_export_path+"/LLM_wos_abstract_topics_metadata_200 - Agreement.csv")

ann.Final_Ann = ["Algorithm and NLP Tasks" if item == "Algorithm" else item for item in ann.Final_Ann]

ann.head()

ann_dict = dict(zip(ann.Topic,ann.Final_Ann))

df["Category"] = df.topics.apply(lambda x: ann_dict[x])

df.head()

df.to_csv(data_export_path+"/LLM_wos_abstract_w_topics_years_positions_categories_200.csv",index=False)
