==> 03_LLM_bib_filter_txt.ipynb <==
import pandas as pd
import numpy as np
import os

data_import_path = "/nfs/turbo/hrg/somar/LLM_bib/raw_data"
data_export_path = "/nfs/turbo/hrg/somar/LLM_bib/result_data"
result_path = "/nfs/turbo/hrg/somar/LLM_bib/result_file"

### Import Data

df_metadata = pd.read_csv(data_export_path+"/LLM_wos_abstract_w_topics_years_positions_categories_200.csv")

df_metadata.head()

df_metadata.info()

# read a text file with records separated by empty lines and output the records as a list
def read_records(file_path):
    with open(file_path, 'r') as file:
        content = file.read()
        records = content.split('\n\n')
        return records[:-1] # the last one if "EF"

# example
file_path = data_import_path+"/download_chatgpt_1_500.txt"
records_list = read_records(file_path)
print(records_list[0])


# create selection index
keep_idx = [0 if item=="Remove" else 1 for item in df_metadata.Category]

len(keep_idx)

# read all txt and save the selected ones to a single txt
selected_pubs = []
file_names = ["1_500","501_1000","1001_1500","1501_2000",
              "2001_2500","2501_3000","3001_3500","3501_4000",
              "4001_4500","4501_5000","5001_5500","5501_5752"]

for i in range(len(file_names)):
    file_name = file_names[i]
    file_path = data_import_path+"/download_chatgpt_"+ file_name +".txt"
    records_list = read_records(file_path)
    print(i, len(records_list))
    for j in range(len(records_list)):
        if keep_idx[i*500+j]==1:
            selected_pubs.append(records_list[j])

selected_pubs_str = "\n\n".join(selected_pubs)
selected_pubs_str += "\n\nEF"

with open(data_import_path+"/download_chatgpt_all.txt", "w") as text_file:
    text_file.write(selected_pubs_str)

df_metadata_clean = df_metadata[df_metadata.Category!="Remove"].reset_index(drop=True)
df_metadata_clean.to_csv(data_export_path+"/LLM_wos_abstract_w_topics_years_positions_categories_200_clean.csv",index=False)

df_metadata_clean = pd.read_csv(data_export_path+"/LLM_wos_abstract_w_topics_years_positions_categories_200_clean.csv")

df_metadata_clean.info()

### Or, we can export txt for each of the Category

set(df_metadata.Category)

selected_pubs_Algo = []
selected_pubs_Crit = []
selected_pubs_Infr = []
selected_pubs_MeEn = []
selected_pubs_SoHu = []

file_names = ["1_500","501_1000","1001_1500","1501_2000",
              "2001_2500","2501_3000","3001_3500","3501_4000",
              "4001_4500","4501_5000","5001_5500","5501_5752"]

for i in range(len(file_names)):
    file_name = file_names[i]
    file_path = data_import_path+"/download_chatgpt_"+ file_name +".txt"
    records_list = read_records(file_path)
    print(i, len(records_list))
    for j in range(len(records_list)):
        this_idx = i*500+j
        if df_metadata.Category[this_idx]=='Algorithm and NLP Tasks':
            selected_pubs_Algo.append(records_list[j])
        elif df_metadata.Category[this_idx]=='Critical studies':
            selected_pubs_Crit.append(records_list[j])
        elif df_metadata.Category[this_idx]=='Infrastructure ':
            selected_pubs_Infr.append(records_list[j])
        elif df_metadata.Category[this_idx]=='Medical and Engineering Applications':
            selected_pubs_MeEn.append(records_list[j])
        elif df_metadata.Category[this_idx]=='Social and Humanitarian Applications':
            selected_pubs_SoHu.append(records_list[j])

selected_pubs_Algo_str = "\n\n".join(selected_pubs_Algo)
selected_pubs_Algo_str += "\n\nEF"
with open(data_import_path+"/download_chatgpt_Algo.txt", "w") as text_file:
    text_file.write(selected_pubs_Algo_str)

selected_pubs_Crit_str = "\n\n".join(selected_pubs_Crit)
selected_pubs_Crit_str += "\n\nEF"
with open(data_import_path+"/download_chatgpt_Crit.txt", "w") as text_file:
    text_file.write(selected_pubs_Crit_str)

selected_pubs_Infr_str = "\n\n".join(selected_pubs_Infr)
selected_pubs_Infr_str += "\n\nEF"
with open(data_import_path+"/download_chatgpt_Infr.txt", "w") as text_file:
    text_file.write(selected_pubs_Infr_str)

selected_pubs_MeEn_str = "\n\n".join(selected_pubs_MeEn)
selected_pubs_MeEn_str += "\n\nEF"
with open(data_import_path+"/download_chatgpt_MeEn.txt", "w") as text_file:
    text_file.write(selected_pubs_MeEn_str)

selected_pubs_SoHu_str = "\n\n".join(selected_pubs_SoHu)
selected_pubs_SoHu_str += "\n\nEF"
with open(data_import_path+"/download_chatgpt_SoHu.txt", "w") as text_file:
    text_file.write(selected_pubs_SoHu_str)
